Moved deployment API into public `org.gradle.deployment`
