+++ subprojects/core/src/main/java/org/gradle/deployment/DeploymentRegistry.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015 the original author or authors.
+ * Copyright 2017 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.deployment.internal;
+package org.gradle.deployment;
 
 import net.jcip.annotations.ThreadSafe;
 import org.gradle.api.Incubating;
@@ -52,6 +52,7 @@ public interface DeploymentRegistry {
     /**
      * Behavior when a deployment is out-of-date.
      */
+    @Incubating
     enum ChangeBehavior {
         /**
          * When changes are pending, block the deployment until all changes are incorporated.
