+++ subprojects/core/src/main/java/org/gradle/deployment/internal/DefaultDeploymentRegistry.java
@@ -22,6 +22,8 @@ import org.gradle.BuildResult;
 import org.gradle.api.logging.Logger;
 import org.gradle.api.logging.Logging;
 import org.gradle.api.model.ObjectFactory;
+import org.gradle.deployment.Deployment;
+import org.gradle.deployment.DeploymentHandle;
 import org.gradle.initialization.ContinuousExecutionGate;
 import org.gradle.internal.Cast;
 import org.gradle.internal.concurrent.CompositeStoppable;
