+++ subprojects/core/src/main/java/org/gradle/deployment/internal/DefaultDeploymentStatus.java
@@ -16,6 +16,8 @@
 
 package org.gradle.deployment.internal;
 
+import org.gradle.deployment.Deployment;
+
 class DefaultDeploymentStatus implements Deployment.Status {
     private final boolean changed;
     private final Throwable failure;
