+++ subprojects/core/src/main/java/org/gradle/deployment/internal/DeploymentInternal.java
@@ -16,6 +16,8 @@
 
 package org.gradle.deployment.internal;
 
+import org.gradle.deployment.Deployment;
+
 public interface DeploymentInternal extends Deployment {
     void outOfDate();
     void upToDate(Throwable failure);
